+++ src/components/calendar.tsx
@@ -1,6 +1,5 @@
 import styled from "@emotion/styled";
-import { css } from '@emotion/react';
-
+import { css } from "@emotion/react";
 import type { FC } from "react";
 import { useMemo } from "react";
 import dayjs from "dayjs";
